Link segments to matrix

+++ src/jtimelag/Segment.java
@@ -1,11 +1,11 @@
 package jtimelag;
 
 public class Segment {
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
     boolean p1Selected;
     boolean p2Selected;
      
-    public Segment(int x1,int y1,int x2,int y2, boolean p1Selected, boolean p2Selected){
+    public Segment(double x1,double y1,double x2,double y2, boolean p1Selected, boolean p2Selected){
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
